Offset changed to byte for efficiency

+++ Project1/dataOutput/address.java
@@ -1,28 +1,30 @@
 package dataOutput;
 
 public class address {
+    //Block ID of the record
     private int blockid;
-    private int offset;
+    //Index of the record in the block
+    private byte offset;
 
     public address(int blockid, int offset){
         // 7 bytes to represent block id
-        this.blockid=blockid;
+        this.blockid = blockid;
         // 1 byte to represent offset since offset is less than 16
-        this.offset=offset;
+        this.offset= (byte) offset;
     }
 
     public void setBlockId(int blockid){
         this.blockid=blockid;
     }
     public void setOffset(int offset){
-        this.offset=offset;
+        this.offset= (byte) offset;
     }
 
     public int getBlockId(){
         return this.blockid;
     }
     public int getOffset(){
-        return this.offset;
+        return (int) this.offset;
     }
 
 }
